Network join and leave start, invite player to plot start.
Added online_users table to database, this is to keep track of the players currently online in the network, since players can be on different servers but should still be grouped as one.
A player disconnects from the network is they are on no server for 2 seconds, this value may be changed based on testing, and how long players take to connect between servers.

Started implementing plot inviting, online players can be invited, they will receive an invite in chat, which they can click on the join the plot. The click event will run a command checking if the player has been invited, and then run the plot adding procedure.
Invited are cancelled if either the player invited disconnects from the network or the plot owner disconnects from the network.