Started Lobby
The Lobby will be completely separate from the rest of the plugin since it'll only run on a single server, the lobby server. By doing this we make sure it'll only run on there and not have any impact on server performance elsewhere.

Started adding portal functionality. Portals will be read from portals.yml with info on how to add portals in the wiki. Portals can only be edited via that file but can be reloaded ingame as to prevent a server restart.