Completed Portal functionality.
portals.yml structure explained in the github wiki page linked in the file.
Portals can be reloaded at any time.
Portals can run commands and custom events. Event functions have not yet been added, but commands should work already.
Checks if players are in the portal every second, this may be altered to be on a shorter interval.