Added /discord link and /discord unlink.
Please received 6 digit code which has to be dm'd to the UK Bot for the account to be linked, this has a 5 minute timeout.

Upon linking the building-related roles are synced with only the highest role being given.

Linking is updated every minute.

Unlinking will also remove the roles in discord.