Bugfixes:
When setting a region to open it was not adding the role to the worldguard region, and when removing it was not removing them.
The gui did not refresh when toggling open/closed status.
When setting an open region to locked it will not remove the jrbuilder role.
Fixed the item not displaying in the build gui when the region was open.